stop excessive content update
